Reuse the context when appropriate
This commit fixes a bug in the `runner::run` function. When the new
image name computed by `cache::image_name` is identical to that of the
image from the given context, the `runner::run` function would try to
construct a new context (as it normally does) with the same image. The
problem is that the original context might delete the image when it's
dropped at the end of the function. The solution for that situation is
to return the original context rather than constructing a new one for
the same image.