Implements ability to change out Docker CLI
This commit adds the ability to swap out the Docker CLI binary for
other drop-in replacements via both a new configuration option
`docker_cli` and the CLI `--docker-cli`.

In the configuration file `docker_cli: string` can be used to set the
Docker CLI binary, and defaults to `docker`.

In a similar manner the CLI option `--docker-cli string` can be used to
set the Docker CLI binary or override the configuration file.

At this time, only drop-in replacements are supported.

Closes #387