implements ability to hide tasks from --list
This commit changes the behavior of `--list` by only showing tasks that
have a description (i.e. `description: string`). The help output of
`--list` is also changed to reflect this new behavior.

If a task requires a description for documentation purposes, but is not
intended to be "public" it is recommended to use a comment instead of
the `description` field.

Closes #388